+++ src/main/java/com/projectboated/backend/web/kanban/kanbanlane/KanbanLaneController.java
@@ -4,7 +4,7 @@ import com.projectboated.backend.domain.account.account.entity.Account;
 import com.projectboated.backend.domain.kanban.kanbanlane.entity.CustomKanbanLane;
 import com.projectboated.backend.domain.kanban.kanbanlane.entity.KanbanLane;
 import com.projectboated.backend.domain.kanban.kanbanlane.service.KanbanLaneService;
-import com.projectboated.backend.web.kanban.kanbanlane.dto.CreateKanbanLaneRequest;
+import com.projectboated.backend.web.kanban.kanbanlane.dto.request.CreateKanbanLaneRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
