+++ src/main/java/com/projectboated/backend/web/task/controller/TaskController.java
@@ -4,8 +4,8 @@ import com.projectboated.backend.domain.account.account.entity.Account;
 import lombok.RequiredArgsConstructor;
 import com.projectboated.backend.domain.task.entity.Task;
 import com.projectboated.backend.domain.task.service.TaskService;
-import com.projectboated.backend.web.task.dto.CreateTaskRequest;
-import com.projectboated.backend.web.task.dto.CreateTaskResponse;
+import com.projectboated.backend.web.task.dto.request.CreateTaskRequest;
+import com.projectboated.backend.web.task.dto.response.CreateTaskResponse;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
