--- src/main/java/com/projectboated/backend/web/task/dto/GetMyTasksResponse.java
@@ -1,4 +1,0 @@
-package com.projectboated.backend.web.task.dto;
-
-public class GetMyTasksResponse {
-}
