+++ src/main/java/com/projectboated/backend/web/task/dto/common/TaskResponse.java
@@ -1,0 +1,30 @@
+package com.projectboated.backend.web.task.dto.common;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.projectboated.backend.domain.task.entity.Task;
+import lombok.Getter;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
+
+@Getter
+public class TaskResponse {
+
+    private Long id;
+    private String name;
+    private String description;
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime deadline;
+    private Long dday;
+    private int fileCount;
+
+    public TaskResponse(Task task) {
+        this.id = task.getId();
+        this.name = task.getName();
+        this.description = task.getDescription();
+        this.deadline = task.getDeadline();
+        this.dday = deadline!=null ? ChronoUnit.DAYS.between(deadline.toLocalDate(), LocalDate.now()) : null;
+        this.fileCount = 0;
+    }
+}
