+++ src/main/java/com/projectboated/backend/web/task/dto/response/GetMyTasksResponse.java
@@ -1,0 +1,4 @@
+package com.projectboated.backend.web.task.dto.response;
+
+public class GetMyTasksResponse {
+}
